% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagemap.R
\name{pagemap-shiny}
\alias{pagemap-shiny}
\alias{pagemapOutput}
\alias{renderPagemap}
\title{Shiny bindings for pagemap}
\usage{
pagemapOutput(outputId)

renderPagemap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a pagemap}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using pagemap within Shiny
applications and interactive Rmd documents.
}
