% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagemap.R
\name{pagemap}
\alias{pagemap}
\title{Mini Map of Page}
\usage{
pagemap(id, box_style = list(), ..., elementId = NULL)
}
\arguments{
\item{id}{\code{string} id for canvas mini map.}

\item{box_style}{a \code{list} of css propery of mini map box.}

\item{...}{additional options passed to \code{pagemap}.}

\item{elementId}{\code{string} id as a valid \code{CSS} element id for htmlwidgets.}
}
\description{
Create mini map for single web page.
}
\examples{
library(pagemap)
pagemap(id = "mini_map")
}
